cli/requirements.txt
click>=8.1.0
requests>=2.31.0
pydantic>=2.5.0
pyyaml>=6.0.1
colorama>=0.4.6
rich>=13.7.0
